Adding a limit to the subquery when converting EXISTS/NOT EXISTS subquery to an aggregate
Prior to this CL, we converted an EXISTS and NOT EXISTS subquery
into an aggregate subquery, and checked the count values (count > 0 for
EXISTS, and count == 0 for NOT EXISTS).

In this CL, I merely added a limit below the count, since all we need to
check is the existence / non existence of a single tuple on the inner
side.