GPORCA incorrectly collapsing FOJ with condition false into a contradiction that will not return any rows
Prior to this fix, the logic that calculated max cardinality for each
logical operator assumed that the FOJ with condition false will always
return empty set. This was used by the following preprocessing step

CExpressionPreprocessor::PexprPruneEmptySubtrees

to eliminate the FOJ subtrees (since it thought had zero output cardinality),
replacing them with a const table get and zero tuples

This collapsing is incorrect. In this CL, the max cardinality logic has been
fixed to ensure that GPORCA plan generates correct plan.